+++ components/lua_ui/widget.cpp
@@ -565,12 +565,12 @@ namespace LuaUi
         });
     }
 
-    void WidgetExtension::mouseWheel(MyGUI::Widget*, int left, int top, float x, float y)
+    void WidgetExtension::mouseWheel(MyGUI::Widget*, float left, float top, float x, float y)
     {
         protectedCall([=, this](LuaUtil::LuaView& view) {
             propagateEvent("mouseWheel", [&](auto w) {
                 sol::table event = view.newTable();
-                event["position"] = osg::Vec2f(static_cast<float>(left), static_cast<float>(top));
+                event["position"] = osg::Vec2f(left, top);
                 event["delta"] = osg::Vec2f(x, y);
                 return event;
             });
